Enforce Qdrant startup deadlines
Bound each liveness probe and retry delay by the startup deadline, and
return the latest transient error once that deadline expires.

Keep API-key plaintext within the adapter boundary, remove unused daemon
runtime dependencies, and align the liveness documentation and delivery
record with the completed behaviour.
