+++ crates/repovec-core/tests/ui/qdrant_api_key_is_opaque.rs
@@ -4,5 +4,5 @@ use repovec_core::appliance::qdrant_liveness::QdrantApiKey;
 
 fn main() {
     let key = QdrantApiKey::parse("qdrant-api-key").expect("test key should parse");
-    let _secret = key.0;
+    let _secret = key.as_secret();
 }
