+++ crates/repovec-mcpd/Cargo.toml
@@ -8,8 +8,6 @@ publish.workspace = true
 
 [dependencies]
 repovec-core = { path = "../repovec-core" }
-tokio.workspace = true
-tracing.workspace = true
 tracing-subscriber.workspace = true
 
 [dev-dependencies]
