+++ crates/repovecd/Cargo.toml
@@ -16,7 +16,6 @@ reqwest = { version = "0.12", default-features = false, features = ["blocking", 
 serde.workspace = true
 tempfile.workspace = true
 thiserror.workspace = true
-tokio.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
